C# - Atualizações - Consumo

+++ GiovanellisPizzariaEntrega1S/TCM - Pizzaria - Semestre 1 - 2014/PCI/Pizzaria/ClassLibrary1/Model/clsConsumo.cs
@@ -11,6 +11,6 @@ namespace DAL.Model
         public int CodProdutoInsumo {get; set;}
         public int CodInsumo {get; set;}
         public int CodProduto {get; set;}
-        public float Quantidade {get; set;}
+        public decimal Quantidade {get; set;}
     }
 }
